refactor(api): moved http api functions to separate namespace
