+++ packages/artalk/src/comment/actions.ts
@@ -16,7 +16,7 @@ export default class CommentActions {
   public vote(type: 'up'|'down') {
     const actionBtn = (type === 'up') ? this.comment.getRender().voteBtnUp : this.comment.getRender().voteBtnDown
 
-    this.ctx.getApi().vote(this.data.id, `comment_${type}`)
+    this.ctx.getApi().comment.vote(this.data.id, `comment_${type}`)
     .then((v) => {
       this.data.vote_up = v.up
       this.data.vote_down = v.down
@@ -45,7 +45,7 @@ export default class CommentActions {
       modify.is_pinned = !modify.is_pinned
     }
 
-    this.ctx.getApi().commentEdit(modify).then((data) => {
+    this.ctx.getApi().comment.commentEdit(modify).then((data) => {
       btnElem.setLoading(false)
 
       // 刷新当前 Comment UI
@@ -64,7 +64,7 @@ export default class CommentActions {
     if (btnElem.isLoading) return // 若正在删除中
 
     btnElem.setLoading(true, `${this.ctx.$t('deleting')}...`)
-    this.ctx.getApi().commentDel(this.data.id, this.data.site_name)
+    this.ctx.getApi().comment.commentDel(this.data.id, this.data.site_name)
       .then(() => {
         btnElem.setLoading(false)
         if (this.cConf.onDelete) this.cConf.onDelete(this.comment)
