+++ packages/artalk/src/list/list.ts
@@ -143,7 +143,7 @@ export default class List extends ListLite {
     // 已阅 API
     const notifyKey = Utils.getQueryParam('atk_notify_key')
     if (notifyKey) {
-      this.ctx.getApi().markRead(notifyKey)
+      this.ctx.getApi().user.markRead(notifyKey)
         .then(() => {
           this.unread = this.unread.filter(o => o.comment_id !== commentId)
           this.updateUnread(this.unread)
@@ -175,7 +175,7 @@ export default class List extends ListLite {
     if (!this.data || !this.data.page) return
 
     this.ctx.editorShowLoading()
-    this.ctx.getApi().pageEdit(this.data.page)
+    this.ctx.getApi().page.pageEdit(this.data.page)
       .then((page) => {
         if (this.data)
           this.data.page = { ...page }
